refactor(services): move adapters into subdirectories and rename to .adapter
Move VCKit DID adapter into adapters/vckit/ and AES-GCM encryption
adapter into adapters/aes-gcm/. Rename VCKitDidService to
VCKitDidAdapter for consistency with the adapter pattern.
