+++ packages/services/src/did-manager/adapters/vckit/vckit-did.schema.ts
