+++ packages/services/src/index.ts
@@ -28,13 +28,13 @@ export type {
 export * from './did-manager/types.js';
 export { didDocumentSchema, verificationMethodSchema } from './did-manager/schemas.js';
 export { didWebToUrl, parseDidMethod, normaliseDidAlias } from './did-manager/utils.js';
-export { VCKitDidService } from './did-manager/adapters/vckit-did.service.js';
+export { VCKitDidAdapter } from './did-manager/adapters/vckit/vckit-did.adapter.js';
 export { verifyDid } from './did-manager/verify.js';
 export type { VerifyDidOptions } from './did-manager/verify.js';
 
 // Encryption
 export type { IEncryptionService } from './encryption/encryption.interface.js';
-export { AesGcmEncryptionAdapter } from './encryption/adapters/aes-gcm.adapter.js';
+export { AesGcmEncryptionAdapter } from './encryption/adapters/aes-gcm/aes-gcm.adapter.js';
 
 // Registry
 export { ServiceType, AdapterType } from './registry/types.js';
@@ -42,5 +42,5 @@ export type { AdapterRegistryEntry, AdapterRegistry } from './registry/types.js'
 export { adapterRegistry } from './registry/registry.js';
 
 // Config schemas
-export { vckitDidConfigSchema, vckitDidSensitiveFields } from './did-manager/adapters/vckit-did.schema.js';
-export type { VCKitDidConfig } from './did-manager/adapters/vckit-did.schema.js';
+export { vckitDidConfigSchema, vckitDidSensitiveFields } from './did-manager/adapters/vckit/vckit-did.schema.js';
+export type { VCKitDidConfig } from './did-manager/adapters/vckit/vckit-did.schema.js';
