+++ packages/services/src/registry/registry.ts
@@ -1,5 +1,5 @@
 import { ServiceType, AdapterType } from './types.js';
-import { vckitDidRegistryEntry } from '../did-manager/adapters/vckit-did.service.js';
+import { vckitDidRegistryEntry } from '../did-manager/adapters/vckit/vckit-did.adapter.js';
 
 export const adapterRegistry = {
   [ServiceType.DID]: {
